4.21

+++ part4/blogList/controllers/blogs.js
@@ -51,8 +51,23 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }else {
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+      if(blog.user.toString() === decodedToken.id.toString()){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } else {
+        response.status(404).end()
+      }
+    } else {
+      response.status(404).end()
+    }
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
